requirements.txt
# TCGP Bot - Requirements
# Python 3.10+

# Discord API
discord.py==2.3.2

# Environment Variables
python-dotenv==1.0.0

# Optional: For enhanced logging
aiohttp==3.9.1

# Optional: For database support (if needed in future)
# sqlalchemy==2.0.23
